deletion checking

+++ app/Http/Controllers/Admin/AdminJobExperienceController.php
@@ -57,6 +57,11 @@ class AdminJobExperienceController extends Controller
 
     public function delete($id)
     {
+        $check = Job::where('job_experience_id',$id)->count();
+        if($check>0) {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+        }
+
         JobExperience::where('id',$id)->delete();
         return redirect()->route('admin_job_experience')->with('success', 'Data is deleted successfully.');
     }
