+++ app/Http/Controllers/Admin/AdminJobLocationController.php
@@ -57,6 +57,11 @@ class AdminJobLocationController extends Controller
 
     public function delete($id)
     {
+        $check = Job::where('job_location_id',$id)->count();
+        if($check>0) {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+        }
+
         JobLocation::where('id',$id)->delete();
         return redirect()->route('admin_job_location')->with('success', 'Data is deleted successfully.');
     }
